+++ src/Bridge/Sylius/config/services.php
@@ -2,10 +2,10 @@
 
 namespace Symfony\Component\DependencyInjection\Loader\Configurator;
 
+use JoliCode\MediaBundle\Bridge\Form\DataTransformer\MediaTransformer;
+use JoliCode\MediaBundle\Bridge\Security\Voter\MediaVoter;
 use JoliCode\MediaBundle\Bridge\Sylius\Admin\Controller\MediaAdminController;
-use JoliCode\MediaBundle\Bridge\Sylius\Admin\Form\DataTransformer\MediaTransformer;
 use JoliCode\MediaBundle\Bridge\Sylius\Admin\Form\Type\MediaChoiceType;
-use JoliCode\MediaBundle\Bridge\Sylius\Admin\Form\Type\UploadType;
 use JoliCode\MediaBundle\Bridge\Sylius\Admin\Grid\MediaGrid;
 use JoliCode\MediaBundle\Bridge\Sylius\Admin\Grid\Provider\MediaGridProvider;
 use JoliCode\MediaBundle\Bridge\Sylius\Asset\Package;
@@ -26,10 +26,10 @@ return static function (ContainerConfigurator $container): void {
     $services
         ->set('joli_media_sylius_admin.config', Config::class)
         ->args([
-            '$visibility' => param('joli_media_easy_admin.visibility'),
-            '$acceptedFiles' => param('joli_media_easy_admin.upload.accepted_files'),
-            '$maxFiles' => param('joli_media_easy_admin.upload.max_files'),
-            '$maxFileSize' => param('joli_media_easy_admin.upload.max_file_size'),
+            '$visibility' => param('joli_media_sylius_admin.visibility'),
+            '$acceptedFiles' => param('joli_media_sylius_admin.upload.accepted_files'),
+            '$maxFiles' => param('joli_media_sylius_admin.upload.max_files'),
+            '$maxFileSize' => param('joli_media_sylius_admin.upload.max_file_size'),
             '$translator' => service('translator')->ignoreOnInvalid(),
         ])
     ;
@@ -69,7 +69,7 @@ return static function (ContainerConfigurator $container): void {
         ->tag('sylius.grid_data_provider')
     ;
 
-    $services->set('joli_media_sylius_admin.form.media_transformer', MediaTransformer::class)
+    $services->set('joli_media_admin.form.data_transformer.media', MediaTransformer::class)
         ->args([
             service('joli_media.resolver'),
         ])
@@ -79,16 +79,14 @@ return static function (ContainerConfigurator $container): void {
         ->args([
             service('joli_media.resolver'),
             service('joli_media.library_container'),
-            service('joli_media_sylius_admin.form.media_transformer'),
+            service('joli_media_admin.form.data_transformer.media'),
         ])
         ->tag('form.type')
     ;
 
-    $services->set('joli_media_sylius_admin.form.upload', UploadType::class)
-        ->args([
-            service('joli_media_sylius_admin.config'),
-        ])
-        ->tag('form.type')
+    // security
+    $services->set('joli_media_admin.security.voter', MediaVoter::class)
+        ->tag('security.voter')
     ;
 
     // twig