+++ src/Bridge/src/Form/Type/DeleteType.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace JoliCode\MediaBundle\Bridge\SonataAdmin\Form\Type;
+namespace JoliCode\MediaBundle\Bridge\Form\Type;
 
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\Extension\Core\Type\HiddenType;
@@ -10,17 +10,11 @@ use Symfony\Component\Validator\Constraints\NotBlank;
 
 class DeleteType extends AbstractType
 {
-    public function getName(): string
-    {
-        return 'joli_media_sonata_admin_delete';
-    }
-
     #[\Override]
     public function configureOptions(OptionsResolver $resolver): void
     {
         $resolver->setDefaults([
             'path' => '',
-            'translation_domain' => 'JoliMediaSonataAdminBundle',
         ]);
     }