+++ src/Bridge/src/Form/Type/RenameType.php
@@ -1,22 +1,16 @@
 <?php
 
-namespace JoliCode\MediaBundle\Bridge\SonataAdmin\Form\Type;
+namespace JoliCode\MediaBundle\Bridge\Form\Type;
 
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\Extension\Core\Type\SubmitType;
 use Symfony\Component\Form\Extension\Core\Type\TextType;
 use Symfony\Component\Form\FormBuilderInterface;
-use Symfony\Component\OptionsResolver\OptionsResolver;
 use Symfony\Component\Validator\Constraints\Length;
 use Symfony\Component\Validator\Constraints\NotBlank;
 
 class RenameType extends AbstractType
 {
-    public function getName(): string
-    {
-        return 'joli_media_sonata_admin_rename';
-    }
-
     #[\Override]
     public function buildForm(FormBuilderInterface $builder, array $options): void
     {
@@ -42,13 +36,5 @@ class RenameType extends AbstractType
                 'label' => 'action.rename_file',
             ])
         ;
-    }
-
-    #[\Override]
-    public function configureOptions(OptionsResolver $resolver): void
-    {
-        $resolver->setDefaults([
-            'translation_domain' => 'JoliMediaSonataAdminBundle',
-        ]);
     }
 }